fix: add discover-peer command
