+++ applications/tari_app_utilities/src/utilities.rs
@@ -205,6 +205,7 @@ pub fn either_to_node_id(either: Either<CommsPublicKey, NodeId>) -> NodeId {
     }
 }
 
+#[derive(Debug)]
 pub struct UniPublicKey(PublicKey);
 
 impl FromStr for UniPublicKey {
