+++ applications/tari_base_node/src/commands/command/mod.rs
@@ -3,6 +3,7 @@ mod block_timing;
 mod check_db;
 mod check_for_updates;
 mod dial_peer;
+mod discover_peer;
 mod get_block;
 mod get_chain_metadata;
 mod get_db_stats;
@@ -82,7 +83,7 @@ pub enum Command {
     BlockTiming(block_timing::Args),
     CalcTiming(block_timing::Args),
     ListReorgs(list_reorgs::Args),
-    // DiscoverPeer,
+    DiscoverPeer(discover_peer::Args),
     GetBlock(get_block::Args),
     SearchUtxo(search_utxo::Args),
     SearchKernel(search_kernel::Args),
@@ -165,6 +166,7 @@ impl HandleCommand<Command> for CommandContext {
             Command::CheckDb(args) => self.handle_command(args).await,
             Command::BlockTiming(args) | Command::CalcTiming(args) => self.handle_command(args).await,
             Command::ListReorgs(args) => self.handle_command(args).await,
+            Command::DiscoverPeer(args) => self.handle_command(args).await,
             Command::GetBlock(args) => self.handle_command(args).await,
             Command::SearchUtxo(args) => self.handle_command(args).await,
             Command::SearchKernel(args) => self.handle_command(args).await,
