+++ applications/tari_base_node/src/commands/parser.rs
@@ -41,7 +41,6 @@ pub enum BaseNodeCommand {
     GetPeer,
     PeriodStats,
     HeaderStats,
-    DiscoverPeer,
     Quit,
     Exit,
 }
